requirements.txt
streamlit
groq
pandas
python-dotenv


# if "GROQ_API_KEY" in st.secrets:
#     GROQ_API_KEY = st.secrets["GROQ_API_KEY"]
# else:
#     GROQ_API_KEY = os.getenv("GROQ_API_KEY")
GROQ_API_KEY = os.getenv("GROQ_API_KEY")

if not GROQ_API_KEY:
    st.error("API Key missing.")
    st.stop()

client = Groq(api_key=GROQ_API_KEY)